+++ packages/@dicebear/avatars-bottts-sprites/src/mouth/square-02.ts
@@ -1,3 +1,0 @@
-export default () => `
-    <rect x="16" y="8" width="44" height="4" rx="2" fill="black" fill-opacity="0.8"/>
-`;