+++ packages/@dicebear/bottts-neutral/src/components/index.ts
@@ -1,0 +1,2 @@
+export { mouth } from './mouth';
+export { eyes } from './eyes';