+++ packages/@dicebear/bottts/src/colors/base.ts
@@ -1,0 +1,23 @@
+import type { ColorGroup } from "../static-types";
+
+export const base: ColorGroup = {
+  'amber': 'rgba(255, 179, 0, 1)',
+  'blue': 'rgba(30, 136, 229, 1)',
+  'blueGrey': 'rgba(84, 110, 122, 1)',
+  'brown': 'rgba(109, 76, 65, 1)',
+  'cyan': 'rgba(0, 172, 193, 1)',
+  'deepOrange': 'rgba(244, 81, 30, 1)',
+  'deepPurple': 'rgba(94, 53, 177, 1)',
+  'green': 'rgba(67, 160, 71, 1)',
+  'grey': 'rgba(117, 117, 117, 1)',
+  'indigo': 'rgba(57, 73, 171, 1)',
+  'lightBlue': 'rgba(3, 155, 229, 1)',
+  'lightGreen': 'rgba(124, 179, 66, 1)',
+  'lime': 'rgba(192, 202, 51, 1)',
+  'orange': 'rgba(251, 140, 0, 1)',
+  'pink': 'rgba(216, 27, 96, 1)',
+  'purple': 'rgba(142, 36, 170, 1)',
+  'red': 'rgba(229, 57, 53, 1)',
+  'teal': 'rgba(0, 137, 123, 1)',
+  'yellow': 'rgba(253, 216, 53, 1)',
+}