+++ packages/@dicebear/bottts/src/index.ts
@@ -1,0 +1,17 @@
+/*!
+ * Bottts (@dicebear/bottts)
+ *
+ * Code licensed under MIT License.
+ * Copyright (c) 2021 Florian Körner
+ *
+ * Design "Bottts" by Pablo Stanley licensed under Free for personal and commercial use..
+ * Source: https://bottts.com/
+ * License: https://bottts.com/
+ */
+
+import { style } from './core';
+
+let { create, meta, schema } = style;
+
+export { create, meta, schema };
+export { Options } from './options';