+++ packages/@dicebear/collection/src/index.ts
@@ -1,6 +1,7 @@
 export * as avataaars from '@dicebear/avataaars';
 export * as avataaarsNeutral from '@dicebear/avataaars-neutral';
-export * as bottts from '@dicebear/avatars-bottts-sprites';
+export * as bottts from '@dicebear/bottts';
+export * as botttsNeutral from '@dicebear/bottts-neutral';
 export * as gridy from '@dicebear/avatars-gridy-sprites';
 export * as identicon from '@dicebear/identicon';
 export * as initials from '@dicebear/initials';